+++ src/components/audio-player/audio-player.e2e.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {mount} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import AudioPlayer from './AudioPlayer.jsx';
+import AudioPlayer from './audio-player.jsx';
 
 Enzyme.configure({adapter: new Adapter()});
