+++ src/components/genre-questionScreen/genre-question-screen.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import GenreQuestionScreen from './GenreQuestionScreen.jsx';
+import GenreQuestionScreen from './genre-question-screen.jsx';
 
 Enzyme.configure({adapter: new Adapter()});
