+++ src/components/genre-questionScreen/genre-question-screen.test.js
@@ -1,17 +1,19 @@
 import React from 'react';
+import GenreQuestionScreen from './genre-question-screen.jsx';
 import ShallowRenderer from 'react-test-renderer/shallow';
-import ArtistQuestionScreen from './ArtistQuestionScreen.jsx';
 
-test(`ArtistQuestionScreen's shapshot`, () => {
+test(`GenreQuestionScreen's snapshot`, () => {
   const renderer = new ShallowRenderer();
   renderer.render(
-      <ArtistQuestionScreen
+      <GenreQuestionScreen
         screenIndex={0}
+        question={{
+          genre: `folk`,
+          answers: [{genre: `rock`, src: `url1`}, {genre: `pop`, src: `url2`}]
+        }}
         onAnswer={jest.fn()}
-        question={{answers: [{}], song: {artist: `str`, src: `str`}}}
       />
   );
-
   const result = renderer.getRenderOutput();
 
   expect(result).toMatchSnapshot();
