+++ src/components/welcome/welcome.e2e.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Welcome from './Welcome';
+import Welcome from './welcome.jsx';
 
 Enzyme.configure({adapter: new Adapter()});
